Fix MSVC warnings in automated-tests. (2/2)

+++ automated-tests/src/dali/utc-Dali-TouchProcessing.cpp
@@ -1179,7 +1179,7 @@ int UtcDaliTouchEventMultipleRenderTasks(void)
   scene.Add(actor);
 
   // Create render task
-  Viewport   viewport(sceneSize.width * 0.5f, sceneSize.height * 0.5f, sceneSize.width * 0.5f, sceneSize.height * 0.5f);
+  Viewport   viewport(static_cast<int>(sceneSize.width * 0.5f), static_cast<int>(sceneSize.height * 0.5f), static_cast<int>(sceneSize.width * 0.5f), static_cast<int>(sceneSize.height * 0.5f));
   RenderTask renderTask(application.GetScene().GetRenderTaskList().CreateTask());
   renderTask.SetViewport(viewport);
   renderTask.SetInputEnabled(true);
@@ -1228,7 +1228,7 @@ int UtcDaliTouchEventMultipleRenderTasksWithChildLayer(void)
   actor.Add(layer);
 
   // Create render task
-  Viewport   viewport(sceneSize.width * 0.5f, sceneSize.height * 0.5f, sceneSize.width * 0.5f, sceneSize.height * 0.5f);
+  Viewport   viewport(static_cast<int>(sceneSize.width * 0.5f), static_cast<int>(sceneSize.height * 0.5f), static_cast<int>(sceneSize.width * 0.5f), static_cast<int>(sceneSize.height * 0.5f));
   RenderTask renderTask(application.GetScene().GetRenderTaskList().CreateTask());
   renderTask.SetViewport(viewport);
   renderTask.SetInputEnabled(true);
@@ -1269,7 +1269,7 @@ int UtcDaliTouchEventOffscreenRenderTasks(void)
   Vector2                  sceneSize(scene.GetSize());
 
   // FrameBufferImage for offscreen RenderTask
-  FrameBuffer frameBuffer = FrameBuffer::New(sceneSize.width, sceneSize.height);
+  FrameBuffer frameBuffer = FrameBuffer::New(static_cast<uint32_t>(sceneSize.width), static_cast<uint32_t>(sceneSize.height));
 
   // Create a renderable actor to display the FrameBufferImage
   Actor renderableActor = CreateRenderableActor(frameBuffer.GetColorTexture());
